example-lock/pycardano/pyproject.toml
[project]
name = "badges-pycardano"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pycardano>=0.13.2",
    "python-dotenv>=1.1.0",
    "urllib3>=2.4.0",
]
